Autenticação de Usuário + CRUD's prontos e funcionais

+++ Bym.UI/Models/Domain/Reserva.cs
@@ -8,7 +8,7 @@ namespace Bym.UI.Models.Domain
     public class Reserva : EntidadeDominio
     {
         public DateTime DataHora { get; set; }
-        public decimal HorasReservadas { get; set; }
+        public int HorasReservadas { get; set; }
         public Sala Sala { get; set; }
         public Usuario Usuario { get; set; }
     }
